add test for DecodeArray, move core runtime calls to different file
